refactor:fixed routing and imports of file as well as file structure
