+++ src/app/Assignment-1/Users/page.jsx
@@ -1,4 +1,5 @@
 import UserCard from "../Component/UserCard";
+
 const users = [
   {
     id: 1,
