+++ src/app/Assignment-1/page.module.css
@@ -1,0 +1,8 @@
+
+.main {
+  display: flex;
+  flex-direction: column;
+  margin-left: 20%;
+  margin-right: 20%;
+
+}
