+++ src/app/page.jsx
@@ -2,10 +2,21 @@
 //Import and render the Greeting component in the App component.
 
 "use client";
-import Greeting from "./Component/Greeting";
-import styles from "@/app/page.module.css"
+
+import styles from "@/app/page.module.css";
 import { useRouter } from "next/navigation";
 export default function Home() {
   const router = useRouter();
-  
+  return (
+    <div className={styles["main"]}>
+      <h2>Assingnment-1</h2>
+      <button
+        onClick={() => {
+          router.push("Assignment-1");
+        }}
+      >
+        Click me
+      </button>
+    </div>
+  );
 }
